+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 Drone Cookbook Changelog
 =========================
 
+v1.0.0
+------
+* BREAKING CHANGE: Use drone containers rather then installing drone package by default
+* Add ability to deploy drone using drone docker container
+
+v0.7.7
+------
+* Uses docker version from attributes (default 1.8.3)
+* Specify docker tls certs or use self-generated
+
+v0.7.0
+------
+* Deploy Drone v0.4.0 (a dev branch)
+* Add dronerc template
+* Add docker recipe to install docker from cookbook
+* Add update recipe to update docker images
+* For integrations generates docker self-signed certs
+
 v0.6.0
 ------
 * Docker can now be disabled (Graham Weldon)